Move export test fixtures
