+++ test/fixtures/export.js
